requirements.txt
alembic==1.11.3
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==1.10.14
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.27.0
typing_extensions==4.10.0
uvicorn==0.29.0
